code reorg and re-enable localstack
