+++ Makefile
@@ -20,11 +20,11 @@ install:
 
 .ONESHELL:
 start-deps:
-	docker-compose up -d postgres rabbit
-	# aws --endpoint-url=http://localhost:4566 s3 mb s3://${BUCKET_STATIC}
-	# aws --endpoint-url=http://localhost:4566 s3 mb s3://${BUCKET_UPLOADS}
-	# aws --endpoint-url=http://localhost:4566 s3 mb s3://${BUCKET_MEDIA_FILES}
-	# aws --endpoint-url=http://localhost:4566 s3 mb s3://${BUCKET_MEDIA_FILE_THUMBNAILS}
+	docker-compose up -d postgres rabbit localstack
+	aws --endpoint-url=http://localhost:4566 s3 mb s3://${BUCKET_STATIC}
+	aws --endpoint-url=http://localhost:4566 s3 mb s3://${BUCKET_UPLOADS}
+	aws --endpoint-url=http://localhost:4566 s3 mb s3://${BUCKET_MEDIA_FILES}
+	aws --endpoint-url=http://localhost:4566 s3 mb s3://${BUCKET_MEDIA_FILE_THUMBNAILS}
 
 .ONESHELL:
 .PHONY: run
