+++ veems/settings.py
@@ -119,7 +119,7 @@ AWS_REGION = os.environ['AWS_REGION']
 AWS_ACCESS_KEY_ID = os.environ['AWS_ACCESS_KEY_ID']
 AWS_SECRET_ACCESS_KEY = os.environ['AWS_SECRET_ACCESS_KEY']
 AWS_S3_USE_SSL = False
-AWS_S3_ENDPOINT_URL = None
+AWS_S3_ENDPOINT_URL = os.environ.get('AWS_S3_ENDPOINT_URL')
 AWS_STORAGE_BUCKET_NAME = os.environ['BUCKET_STATIC']
 BUCKET_UPLOADS = os.environ['BUCKET_UPLOADS']
 BUCKET_MEDIA_FILE_THUMBNAILS = os.environ[
